Fixed a couple of bugs. The callback seems mostly ok. Now I am going to look at what is making things seemingly break

+++ include/runtime/init_close.h
@@ -10,7 +10,7 @@
 
 using namespace std;
 
-#define RING_ELEMENT sizeof(double)
+#define RING_ELEMENT sizeof(float)
 #define RING_SIZE (1048576 / RING_ELEMENT) //1MB ring buffer
 #define SAMPLE_JUMP 10
